FirmwareHelpers/requirements.txt
esptool
pyserial~=3.5